requirements.txt
# requirements.txt
pandas
scikit-learn
dvc[azure]
fastapi
uvicorn[standard]
python-dotenv
joblib
PyYAML
pytest  
pytest-mock 
requests 
